Remove old taskview
